move stuff to matrixeventprocessor and make matrixcommandhandler
