+++ src/bot.ts
@@ -10,7 +10,7 @@ import {
     DiscordMessageProcessorOpts,
     DiscordMessageProcessorMatrixResult,
 } from "./discordmessageprocessor";
-import { MatrixEventProcessor, MatrixEventProcessorOpts } from "./matrixeventprocessor";
+import { MatrixEventProcessor, IMatrixEventProcessorResult } from "./matrixeventprocessor";
 import { PresenceHandler } from "./presencehandler";
 import { Provisioner } from "./provisioner";
 import { UserSyncroniser } from "./usersyncroniser";
@@ -77,13 +77,18 @@ export class DiscordBot {
 
     public setBridge(bridge: Bridge) {
         this.bridge = bridge;
-        this.mxEventProcessor = new MatrixEventProcessor(
-            new MatrixEventProcessorOpts(this.config, bridge, this),
-        );
     }
 
     public setRoomHandler(roomHandler: MatrixRoomHandler) {
         this.roomHandler = roomHandler;
+    }
+
+    public setEventProcessor(eventProcessor: MatrixEventProcessor) {
+        this.mxEventProcessor = eventProcessor;
+    }
+
+    get Provisioner(): Provisioner {
+        return this.provisioner;
     }
 
     get ClientFactory(): DiscordClientFactory {
@@ -96,6 +101,14 @@ export class DiscordBot {
 
     get ChannelSyncroniser(): ChannelSyncroniser {
         return this.channelSync;
+    }
+
+    get RoomHandler(): MatrixRoomHandler {
+        return this.roomHandler;
+    }
+
+    public getBotId(): string {
+        return this.bridge.getIntent().getClient().getUserId();
     }
 
     public GetIntentFromDiscordMember(member: Discord.GuildMember | Discord.User): Intent {
@@ -300,45 +313,23 @@ export class DiscordBot {
         }
     }
 
-    public async ProcessMatrixStateEvent(event: IMatrixEvent): Promise<void> {
-        log.verbose(`Got state event from ${event.room_id} ${event.type}`);
-        const channel = await this.GetChannelFromRoomId(event.room_id) as Discord.TextChannel;
-        const msg = this.mxEventProcessor.StateEventToMessage(event, channel);
+    public async sendAsBot(msg: string, channel: Discord.TextChannel, event: IMatrixEvent): Promise<void> {
         if (!msg) {
             return;
         }
-        let res = await channel.send(msg);
-        if (!Array.isArray(res)) {
-            res = [res];
-        }
-        await Util.AsyncForEach(res, async (m: Discord.Message) => {
-            log.verbose("Sent (state msg) ", m.id);
-            this.sentMessages.push(m.id);
-            const evt = new DbEvent();
-            evt.MatrixId = `${event.event_id};${event.room_id}`;
-            evt.DiscordId = m.id;
-            evt.GuildId = channel.guild.id;
-            evt.ChannelId = channel.id;
-            await this.store.Insert(evt);
-        });
+        const res = await channel.send(msg);
+        await this.StoreMessagesSent(res, channel, event);
     }
 
-    public async ProcessMatrixMsgEvent(event: IMatrixEvent, guildId: string, channelId: string): Promise<void> {
-        const mxClient = this.bridge.getClientFactory().getClientAs();
-        log.verbose(`Looking up ${guildId}_${channelId}`);
-        const result = await this.LookupRoom(guildId, channelId, event.sender);
-        const chan = result.channel;
-        const botUser = result.botUser;
-
-        const embedSet = await this.mxEventProcessor.EventToEmbed(event, chan);
+    public async send(
+        embedSet: IMatrixEventProcessorResult,
+        opts: Discord.MessageOptions,
+        roomLookup: ChannelLookupResult,
+        event: IMatrixEvent,
+    ): Promise<void> {
+        const chan = roomLookup.channel;
+        const botUser = roomLookup.botUser;
         const embed = embedSet.messageEmbed;
-        const opts: Discord.MessageOptions = {};
-        const file = await this.mxEventProcessor.HandleAttachment(event, mxClient);
-        if (typeof(file) === "string") {
-            embed.description += " " + file;
-        } else {
-            opts.file = file;
-        }
 
         let msg: Discord.Message | null | (Discord.Message | null)[] = null;
         let hook: Discord.Webhook | undefined;
@@ -376,24 +367,10 @@ export class DiscordBot {
                 opts.embed = embed;
                 msg = await chan.send("", opts);
             }
+            await this.StoreMessagesSent(msg, chan, event);
         } catch (err) {
             log.error("Couldn't send message. ", err);
         }
-        if (!Array.isArray(msg)) {
-            msg = [msg];
-        }
-        await Util.AsyncForEach(msg, async (m: Discord.Message) => {
-            log.verbose("Sent ", m.id);
-            this.sentMessages.push(m.id);
-            const evt = new DbEvent();
-            evt.MatrixId = `${event.event_id};${event.room_id}`;
-            evt.DiscordId = m.id;
-            // Webhooks don't send guild info.
-            evt.GuildId = guildId;
-            evt.ChannelId = channelId;
-            await this.store.Insert(evt);
-        });
-        return;
     }
 
     public async ProcessMatrixRedact(event: IMatrixEvent) {
@@ -734,5 +711,28 @@ export class DiscordBot {
                 }
             }
         }
+    }
+
+    private async StoreMessagesSent(
+        msg: Discord.Message | null | (Discord.Message | null)[],
+        chan: Discord.TextChannel,
+        event: IMatrixEvent,
+    ) {
+        if (!Array.isArray(msg)) {
+            msg = [msg];
+        }
+        await Util.AsyncForEach(msg, async (m: Discord.Message) => {
+            if (!m) {
+                return;
+            }
+            log.verbose("Sent ", m.id);
+            this.sentMessages.push(m.id);
+            const evt = new DbEvent();
+            evt.MatrixId = `${event.event_id};${event.room_id}`;
+            evt.DiscordId = m.id;
+            evt.GuildId = chan.guild.id;
+            evt.ChannelId = chan.id;
+            await this.store.Insert(evt);
+        });
     }
 }
